qr_gen/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
pypng==0.20220715.0
PyQRCode==1.2.1
sqlparse==0.4.2
